refactor cmd to commands
